Format SQL queries

+++ admin/user_perm.php
@@ -46,11 +46,8 @@ if (isset($_POST['falsify'])
     // if you forbid access to a category, all sub-categories become
     // automatically forbidden
     $subcats = get_subcat_ids($_POST['cat_true']);
-    $query = '
-DELETE FROM user_access
-  WHERE user_id = ' . $page['user'] . '
-    AND cat_id IN (' . implode(',', $subcats) . ')
-;';
+    $subcats_ = implode(',', $subcats);
+    $query = "DELETE FROM user_access WHERE user_id = {$page['user']} AND cat_id IN ({$subcats_});";
     pwg_query($query);
 } elseif (isset($_POST['trueify'])
     and isset($_POST['cat_false'])
@@ -89,15 +86,9 @@ $template->assign(
 // retrieve category ids authorized to the groups the user belongs to
 $group_authorized = [];
 
-$query = '
-SELECT DISTINCT cat_id, c.uppercats, c.global_rank
-  FROM user_group AS ug
-    INNER JOIN group_access AS ga
-      ON ug.group_id = ga.group_id
-    INNER JOIN categories AS c
-      ON c.id = ga.cat_id
-  WHERE ug.user_id = ' . $page['user'] . '
-;';
+$query =
+"SELECT DISTINCT cat_id, c.uppercats, c.global_rank FROM user_group AS ug INNER JOIN group_access AS ga ON ug.group_id = ga.group_id
+ INNER JOIN categories AS c ON c.id = ga.cat_id WHERE ug.user_id = {$page['user']};";
 $result = pwg_query($query);
 
 if (pwg_db_num_rows($result) > 0) {
@@ -117,17 +108,12 @@ if (pwg_db_num_rows($result) > 0) {
 }
 
 // only private categories are listed
-$query_true = '
-SELECT id,name,uppercats,global_rank
-  FROM categories INNER JOIN user_access ON cat_id = id
-  WHERE status = \'private\'
-    AND user_id = ' . $page['user'];
+$query_true = "SELECT id, name, uppercats, global_rank FROM categories INNER JOIN user_access ON cat_id = id WHERE status = 'private' AND user_id = {$page['user']}";
 if (count($group_authorized) > 0) {
-    $query_true .= '
-    AND cat_id NOT IN (' . implode(',', $group_authorized) . ')';
+    $group_authorized_ = implode(',', $group_authorized);
+    $query_true .= " AND cat_id NOT IN ({$group_authorized_})";
 }
-$query_true .= '
-;';
+$query_true .= ';';
 display_select_cat_wrapper($query_true, [], 'category_option_true');
 
 $result = pwg_query($query_true);
@@ -136,20 +122,16 @@ while ($row = pwg_db_fetch_assoc($result)) {
     $authorized_ids[] = $row['id'];
 }
 
-$query_false = '
-SELECT id,name,uppercats,global_rank
-  FROM categories
-  WHERE status = \'private\'';
+$query_false = "SELECT id, name, uppercats, global_rank FROM categories WHERE status = 'private'";
 if (count($authorized_ids) > 0) {
-    $query_false .= '
-    AND id NOT IN (' . implode(',', $authorized_ids) . ')';
+    $authorized_ids_ = implode(',', $authorized_ids);
+    $query_false .= " AND id NOT IN ({$authorized_ids_})";
 }
 if (count($group_authorized) > 0) {
-    $query_false .= '
-    AND id NOT IN (' . implode(',', $group_authorized) . ')';
+    $group_authorized_ = implode(',', $group_authorized);
+    $query_false .= " AND id NOT IN ({$group_authorized_})";
 }
-$query_false .= '
-;';
+$query_false .= ';';
 display_select_cat_wrapper($query_false, [], 'category_option_false');
 
 $template->assign('PWG_TOKEN', get_pwg_token());
